+++ src/reader/_sqlite_utils.py
@@ -18,14 +18,13 @@ from contextlib import closing
 from contextlib import contextmanager
 from contextlib import nullcontext
 from dataclasses import dataclass
-from datetime import datetime
 from typing import Any
 from typing import cast
 from typing import no_type_check
 from typing import TypeVar
 
 
-SQLiteType = TypeVar('SQLiteType', None, int, float, str, bytes, datetime)
+SQLiteType = TypeVar('SQLiteType', None, int, float, str, bytes)
 
 
 @contextmanager